itcurves_deploy/services/survey-service/requirements.txt
fastapi==0.115.12
uvicorn==0.34.2
openai>=2.0.0
httpx>=0.27.0
sqlalchemy>=2.0
psycopg2-binary>=2.9
python-dotenv>=1.0.0
requests>=2.31.0
APScheduler==3.11.0
mailersend>=2.0.0
